major updates to consolidating the admin and cli exporters
